requirements.txt
boto3>=1.40.46,<2.0
python-dotenv>=1.1.1,<2.0
folium>=0.20.0,<1.0
geopandas>=1.1.1,<2.0
numpy>=1.26.4,<3.0
pandas>=2.3.3,<3.0
pyproj>=3.7.2,<4.0
rasterio>=1.4.3,<2.0
requests>=2.32.5,<3.0
rioxarray>=0.19.0,<1.0
shapely>=2.1.2,<3.0
scikit-learn<2.0
xarray>=2025.10.0,<2026.0